renamed some stuff, cleaned up some stuff, updated composer.json and README.md. changed namespace
