+++ README.md
@@ -1,36 +1,30 @@
 # Laravel Model Documenter
 
-## Getting started
-
-`php artisan vendor:publish --provider="Enz0project\ModelDocumenter\ModelDocumenterProvider" --tag="config"`
-
 ## Running the command
 
-`php artisan enz0project:model-documenter` will run the model-documenter on all models specified in the folder in `./config/modeldocumenter.php`. 
+`php artisan model:document` will run the model-documenter on all models in `app_path('Models')` and subfolders.
 
-If you want to run the the model-documenter only on specific models, you can give it the filename without extension as an argument:
-`php artisan enz0project:model-documenter ForumThread` will make it run only on `ForumThread.php` in the models folder.
-If you want to run it on multiple, specific models, you can separate them by comma:
-`php artisan enz0project:model-documenter User,ForumThread,ForumPost,ForumComment`
+If you want to run the command only on specific models, you can give it the filename without extension as an argument:
+`php artisan model:document ForumThread` will make it run only on `ForumThread.php` in the models folder. The argument
+also takes multiple files, separated by comma:
+`php artisan model:document ForumThread,ForumPost`
 
 ## Docs
 
-If you want to use this outside of the command or write your own command, you can use the ModelAnalyzer:
+If you want to use this outside of the command or write your own command, you can use the ModelData directly:
 ```php
 <?php
-    $modelAnalyzer = new ModelAnalyzer();
-    $modelData = $modelAnalyzer->analyze('/path/to/file.php');
+    $modelData = new ModelData('/path/to/file.php');
 ```
 
-The `analyze` method gives back a ModelData object that can be used like this:
+The ModelData has several public properties that get populated by the constructor:
 
 ```php
 <?php
-    string $name = $modelData->getName(); // The models class name
-    int $type = $modelData->getType(); // Is one of the constants in ModelAnalyzer, i.e. TYPE_CLASS or TYPE_INTERFACE
-    array $fileContents = $modelData->getFileContents(); // Contains all lines in the .php file
-    string $classDocBlock = $modelData->getClassDocBlock(); // The old classlevel docblock, if one exists
-    array $properties = $modelData->getProperties(); // An array of properties where the key is the property name and the value is its php type
-    array|null $relations = $modelData->getRelations(); // An array of relations where the key is the relation name and the value is its php type
-    array $requiredImports = $modelData->getRequiredImports(); // An array of imports required to be in the file, i.e. Carbon or Collection
+    string $modelData->name; // The models class name
+    array $modelData->fileContents; // Contains all lines in the .php file
+    string $modelData->classDocBlock; // The old classlevel docblock, if one exists
+    array $modelData->properties; // An array of properties where the key is the property name and the value is its php type
+    array $modelData->relations; // An array of relations where the key is the relation name and the value is its php type
+    array $modelData->requiredImports; // An array of imports required to be in the file, i.e. Carbon or Collection
 ```
