+++ composer.json
@@ -1,27 +1,31 @@
 {
-	"name": "enz0project/laravel-model-documenter",
+	"name": "omniway/laravel-model-documenter",
 	"description": "Adds docblocks to Laravel models",
 	"type": "library",
 	"require": {
-		"Illuminate/Support": ">=6.0",
-		"Illuminate/Console": ">=6.0"
+		"illuminate/support": ">=6.0",
+		"illuminate/console": ">=6.0"
 	},
 	"license": "MIT",
 	"authors": [
 		{
 			"name": "Lars Bergvall",
 			"email": "lars.m.bergvall@gmail.com"
+		},
+		{
+			"name": "Jonas Lindh Morén",
+			"email": "jonas.lindh-moren@academedia.se"
 		}
 	],
 	"autoload": {
 		"psr-4": {
-			"Enz0project\\ModelDocumenter\\": "src/"
+			"Omniway\\ModelDocumenter\\": "src/"
 		}
 	},
 	"extra": {
 		"laravel": {
 			"providers": [
-				"Enz0project\\ModelDocumenter\\ModelDocumenterProvider"
+				"Omniway\\ModelDocumenter\\ModelDocumenterProvider"
 			],
 			"aliases": {
