--- src/Exceptions/NoTableException.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace Enz0project\ModelDocumenter\Exceptions;
-
-class NoTableException extends \Exception {
-
-}
