--- src/config/modeldocumenter.php
@@ -1,5 +1,0 @@
-<?php
-
-return [
-	'modelPath' => 'App',
-];
